Add a rudimentary bus.
The bus accepts messages and queues them, sending them in another
thread. Additionally, it monitors the name server for new services,
registering them with the bus automatically.

The messages are sent to services on the basis of their type; a service
registers itself as listening for messages with a given type, and these
are sent to the services. We're thinking of extending that system, but
this is the way it works right now.